fix: layout enhancement
