+++ platforms/metagram/src/lib/fragments/index.ts
@@ -18,3 +18,4 @@ export { default as SettingsDeleteButton } from './SettingsDeleteButton/Settings
 export { default as SettingsTile } from './SettingsTile/SettingsTile.svelte';
 export { default as UserRequest } from './UserRequest/UserRequest.svelte';
 export { default as UploadedPostView } from './UploadedPostView/UploadedPostView.svelte';
+export { default as AddPostModal } from './AddPostModal/AddPostModal.svelte';
